Rewrite the linux setup script to use flags

+++ setup.sh
@@ -4,14 +4,18 @@
 . ./.env
 
 # Enable the rc.local script.
-sudo cp ./script/linux/rc.local /etc/
-sudo chmod +x /etc/rc.local
-sudo systemctl enable rc-local
-sudo systemctl restart rc-local
+if [ $APP_SETUP_IS_RCLOCAL_ENABLED ]; then
+	sudo cp ./script/linux/rc.local /etc/
+	sudo chmod +x /etc/rc.local
+	sudo systemctl enable rc-local
+	sudo systemctl restart rc-local
+fi
 
-# Remove useless folders.
-rm -rf ~/Public
-rm -rf ~/Templates
+# Remove useless directories.
+if [ $APP_SETUP_IS_REMOVE_USLESS_DIRS_ENABLED ]; then
+	rm -rf ~/Public
+	rm -rf ~/Templates
+fi
 
 # Copy, setup and run the update script.
 cp ./script/linux/update_system.sh ~/
@@ -26,8 +30,15 @@ chmod +x ~/upgrade_release.sh
 # Install the Build Essential package.
 sudo apt-get install build-essential -y
 
+# Install the Python3 package.
+if [ $APP_SETUP_IS_INSTALL_PYTHON3_ENABLED ]; then
+	sudo apt-get install python3 -y
+fi
+
 # Configure python to point to the python 3.x implementation.
-sudo update-alternatives --install /usr/bin/python python /usr/bin/python3 1
+if [ $APP_SETUP_IS_CONFIGURE_PYTHON3_ALTERNATIVE_ENABLED ]; then
+	sudo update-alternatives --install /usr/bin/python python /usr/bin/python3 1
+fi
 
 # Install the pip for python3 package.
 curl https://bootstrap.pypa.io/get-pip.py -o ~/get-pip.py
@@ -47,4 +58,6 @@ npm i
 npm start script/git
 
 # Load the vscode configurations files.
-npm start script/vscode
+if [ $APP_SETUP_IS_VSCODE_CONFIGURATION_ENABLED ]; then
+	npm start script/vscode
+fi
